LayoutTests/http/wpt/webrtc/script-transform-no-options.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
promise_test(async (test) => {
    const worker = new Worker('no-options-transform-worker.js');

    const transform = new RTCRtpScriptTransform(worker);
    assert_equals(await new Promise(resolve => worker.onmessage = event => resolve(event.data)), undefined);
}, "script transform without options");

promise_test(async (test) => {
    const worker = new Worker('no-options-transform-worker.js');

    const transform = new RTCRtpScriptTransform(worker, undefined);
    assert_equals(await new Promise(resolve => worker.onmessage = event => resolve(event.data)), undefined);
}, "script transform options as undefined");
        </script>
    </body>
</html>
